Update type to follow newest tpying spec
